cracking_the_data_engineering_interview/src/requirements.txt
azure-storage-blob==12.26.0
beautifulsoup4==4.14.3
numpy==2.3.5
pandas==2.3.3
psycopg2-binary==2.9.11
pyarrow==22.0.0
Requests==2.32.5
SQLAlchemy==2.0.45
